requirements.txt
apache-airflow==3.1.8
apache-airflow-providers-fab==3.6.0
apache-airflow-providers-cncf-kubernetes==10.15.0
kubernetes==35.0.0